requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.2
pandas==2.2.3
tldextract==5.1.2
pyyaml==6.0.2
joblib==1.4.2
python-multipart==0.0.9